refactor: react-router added, pages separated
