+++ frontend/package.json
@@ -26,7 +26,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.2.1",
-    "react-router-dom": "^6.17.0",
+    "react-router": "6",
+    "react-router-dom": "^6.24.0",
     "react-toastify": "^10.0.5",
     "react-youtube": "^10.1.0",
     "tailwindcss": "^3.3.3",
