+++ frontend/src/main.tsx
@@ -1,10 +1,13 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import Dashboard from './pages/Dashboard.tsx'
 import './index.css'
+import App from './routes'
+import { BrowserRouter } from 'react-router-dom'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <Dashboard />
+    <BrowserRouter>
+    <App />
+    </BrowserRouter>
   </React.StrictMode>,
 )
