+++ frontend/src/pages/NotFound/index.tsx
@@ -1,0 +1,19 @@
+/* eslint-disable jsx-a11y/anchor-is-valid */
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+const NotFound = () => {
+  return (
+    <div className="min-h-screen bg-gray-100 flex flex-col items-center justify-center">
+      <div className="text-center">
+        <h1 className="text-6xl font-bold text-gray-800">404</h1>
+        <p className="text-xl text-gray-600 mb-4">Page not found</p>
+        <Link to={'/'}>
+          <a className="text-red-600 hover:text-red-800">Back to homepage</a>
+        </Link>
+      </div>
+    </div>
+  );
+};
+
+export default NotFound;
